Fix log rotation
When there are no logs to delete using the LOG_ROTATE variable will cause the container to crash.

Thanks @Skaronator

+++ sync.sh
@@ -8,7 +8,7 @@ echo "INFO: Starting sync.sh pid $$ $(date)"
 if [ ! -z "${ROTATE_LOG##*[!0-9]*}" ]
 then
   echo "INFO: Removing logs older than $ROTATE_LOG day(s)..."
-  find /logs/*.txt -mtime +$ROTATE_LOG -type f -delete
+  test -d /logs/*.txt && find /logs/*.txt -mtime +$ROTATE_LOG -type f -delete
 fi
 
 if [ `lsof | grep $0 | wc -l | tr -d ' '` -gt 1 ]
